feat: structured bottom tab navigation
