+++ frontend/src/components/TabBar.vue
@@ -7,8 +7,8 @@ const route = useRoute()
 
 const items: TabItem[] = [
   { label: 'Home', to: '/home', icon: Home },
-  { label: 'Beispiel', to: '/example', icon: Car },
-  { label: 'Styles', to: '/styles', icon: User }
+  { label: 'Fahrten', to: '/my_rides', icon: Car },
+  { label: 'Profil', to: '/profile', icon: User }
 ]
 
 const isActive = (path: string): boolean => route.path === path
